+++ app/src/main/java/isel/csee/jcctokener/generators/SemanticVectorGenerator.java
@@ -38,7 +38,7 @@ public class SemanticVectorGenerator {
         for(int i = 0; i < jCCNodeList.size(); i++) {
             if(jCCNodeList.get(i).getSemanticType() == type1) {
                 int tempStep = 0;
-                int[] semanticVector = new int[25];
+                double[] semanticVector = new double[25];
 
                 semanticVector = getRelatedValues(jCCNodeList.get(i), tempStep, countStep, semanticVector);
 
@@ -50,7 +50,7 @@ public class SemanticVectorGenerator {
     public void createOperatorSemanticVector() { // type 2에 대한 semantic vector 생성
         for(int i = 0; i < jCCNodeList.size(); i++) {
             if(jCCNodeList.get(i).getSemanticType() == type2) {
-                int[] semanticVector = new int[25];
+                double[] semanticVector = new double[25];
 
                 for(int k = 0; k < jCCNodeList.get(i).getIndexListOfEdges().size(); k++) { // 관련 있는 값들의 semantic vector 모두 더해주기
                     for(int t = 0; t < 25; t++) {
@@ -67,7 +67,7 @@ public class SemanticVectorGenerator {
     public void createMethodSemanticVector() { // type 3에 대한 semantic vector 생성
         for(int i = 0; i < jCCNodeList.size(); i++) {
             if(jCCNodeList.get(i).getSemanticType() == type3) {
-                int[] semanticVector = new int[25];
+                double[] semanticVector = new double[25];
 
                 for(int k = 0; k < jCCNodeList.get(i).getIndexListOfEdges().size(); k++) { // 관련 있는 값들의 semantic vector 모두 더해주기
                     for(int t = 0; t < 25; t++) {
@@ -85,7 +85,7 @@ public class SemanticVectorGenerator {
 
     }
     // 해당 step까지 올라가서 존재하는 모든 variable structure vector 다 가져와서 더해주기
-    public int[] getRelatedValues(jCCNode node, int tempCount, int totalCount, int[] semanticVector) {
+    public double[] getRelatedValues(jCCNode node, int tempCount, int totalCount, double[] semanticVector) {
         for(int i = 0; i < 25; i++) { // 현재 노드의 structure vector 더해주기
             semanticVector[i] += node.getStructureVector()[i];
         }
